requirements.txt
# Web 框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# 数据验证
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# 认证安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
# 使用预编译版本避免需要 Rust
cryptography>=41.0.7

# 缓存与异步任务
redis==5.0.1
celery==5.3.4

# AI 模型 SDK
openai==1.3.0
dashscope==1.14.0
httpx==0.25.2

# 向量数据库
pymilvus==2.3.4

# 文本处理与 Embedding
langchain==0.1.0
langchain-community==0.0.10
sentence-transformers==2.2.2

# 文档解析
pypdf==3.17.0
python-docx==1.1.0
python-pptx==0.6.23
markdown==3.5.1

# 工具
python-dotenv==1.0.0
loguru==0.7.2
tenacity==8.2.3

